+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+## 0.16.0 - 2025-06-12
+
 ***Added:***
 
 - Persist cache directories for the `linux-container` developer environment type